ARGO-2687 move handlers file to package and break it in several files
