+++ routing.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/ARGOeu/argo-messaging/brokers"
 	"github.com/ARGOeu/argo-messaging/config"
+	"github.com/ARGOeu/argo-messaging/handlers"
 	oldPush "github.com/ARGOeu/argo-messaging/push"
 	push "github.com/ARGOeu/argo-messaging/push/grpc/client"
 	"github.com/ARGOeu/argo-messaging/stores"
@@ -44,16 +45,16 @@ func NewRouting(cfg *config.APICfg, brk brokers.Broker, str stores.Store, mgr *o
 		var handler http.HandlerFunc
 		handler = route.Handler
 
-		handler = WrapLog(handler, route.Name)
+		handler = handlers.WrapLog(handler, route.Name)
 
 		// skip authentication/authorization for the health status and profile api calls
 		if route.Name != "ams:healthStatus" && "users:profile" != route.Name && route.Name != "version:list" {
-			handler = WrapAuthorize(handler, route.Name)
-			handler = WrapAuthenticate(handler)
+			handler = handlers.WrapAuthorize(handler, route.Name)
+			handler = handlers.WrapAuthenticate(handler)
 		}
 
-		handler = WrapValidate(handler)
-		handler = WrapConfig(handler, cfg, brk, str, mgr, c)
+		handler = handlers.WrapValidate(handler)
+		handler = handlers.WrapConfig(handler, cfg, brk, str, mgr, c)
 
 		ar.Router.
 			PathPrefix("/v1").
@@ -71,64 +72,64 @@ func NewRouting(cfg *config.APICfg, brk brokers.Broker, str stores.Store, mgr *o
 // Global list populated with default routes
 var defaultRoutes = []APIRoute{
 
-	{"ams:metrics", "GET", "/metrics", OpMetrics},
-	{"ams:healthStatus", "GET", "/status", HealthCheck},
-	{"ams:dailyMessageAverage", "GET", "/metrics/daily-message-average", DailyMessageAverage},
-	{"users:byToken", "GET", "/users:byToken/{token}", UserListByToken},
-	{"users:byUUID", "GET", "/users:byUUID/{uuid}", UserListByUUID},
-	{"users:list", "GET", "/users", UserListAll},
-	{"users:profile", "GET", "/users/profile", UserProfile},
-	{"users:show", "GET", "/users/{user}", UserListOne},
-	{"users:refreshToken", "POST", "/users/{user}:refreshToken", RefreshToken},
-	{"users:create", "POST", "/users/{user}", UserCreate},
-	{"users:update", "PUT", "/users/{user}", UserUpdate},
-	{"users:delete", "DELETE", "/users/{user}", UserDelete},
-	{"registrations:newUser", "POST", "/registrations", RegisterUser},
-	{"registrations:acceptNewUser", "POST", "/registrations/{uuid}:accept", AcceptRegisterUser},
-	{"registrations:declineNewUser", "POST", "/registrations/{uuid}:decline", DeclineRegisterUser},
-	{"registrations:show", "GET", "/registrations/{uuid}", ListOneRegistration},
-	{"registrations:list", "GET", "/registrations", ListAllRegistrations},
-	{"projects:list", "GET", "/projects", ProjectListAll},
-	{"projects:metrics", "GET", "/projects/{project}:metrics", ProjectMetrics},
-	{"projects:addUser", "POST", "/projects/{project}/members/{user}:add", ProjectUserAdd},
-	{"projects:removeUser", "POST", "/projects/{project}/members/{user}:remove", ProjectUserRemove},
-	{"projects:showUser", "GET", "/projects/{project}/members/{user}", ProjectUserListOne},
-	{"projects:createUser", "POST", "/projects/{project}/members/{user}", ProjectUserCreate},
-	{"projects:updateUser", "PUT", "/projects/{project}/members/{user}", ProjectUserUpdate},
-	{"projects:listUsers", "GET", "/projects/{project}/members", ProjectListUsers},
-	{"projects:show", "GET", "/projects/{project}", ProjectListOne},
-	{"projects:create", "POST", "/projects/{project}", ProjectCreate},
-	{"projects:update", "PUT", "/projects/{project}", ProjectUpdate},
-	{"projects:delete", "DELETE", "/projects/{project}", ProjectDelete},
-	{"subscriptions:list", "GET", "/projects/{project}/subscriptions", SubListAll},
-	{"subscriptions:listByTopic", "GET", "/projects/{project}/topics/{topic}/subscriptions", ListSubsByTopic},
-	{"subscriptions:offsets", "GET", "/projects/{project}/subscriptions/{subscription}:offsets", SubGetOffsets},
-	{"subscriptions:timeToOffset", "GET", "/projects/{project}/subscriptions/{subscription}:timeToOffset", SubTimeToOffset},
-	{"subscriptions:acl", "GET", "/projects/{project}/subscriptions/{subscription}:acl", SubACL},
-	{"subscriptions:metrics", "GET", "/projects/{project}/subscriptions/{subscription}:metrics", SubMetrics},
-	{"subscriptions:show", "GET", "/projects/{project}/subscriptions/{subscription}", SubListOne},
-	{"subscriptions:create", "PUT", "/projects/{project}/subscriptions/{subscription}", SubCreate},
-	{"subscriptions:delete", "DELETE", "/projects/{project}/subscriptions/{subscription}", SubDelete},
-	{"subscriptions:pull", "POST", "/projects/{project}/subscriptions/{subscription}:pull", SubPull},
-	{"subscriptions:acknowledge", "POST", "/projects/{project}/subscriptions/{subscription}:acknowledge", SubAck},
-	{"subscriptions:verifyPushEndpoint", "POST", "/projects/{project}/subscriptions/{subscription}:verifyPushEndpoint", SubVerifyPushEndpoint},
-	{"subscriptions:modifyAckDeadline", "POST", "/projects/{project}/subscriptions/{subscription}:modifyAckDeadline", SubModAck},
-	{"subscriptions:modifyPushConfig", "POST", "/projects/{project}/subscriptions/{subscription}:modifyPushConfig", SubModPush},
-	{"subscriptions:modifyOffset", "POST", "/projects/{project}/subscriptions/{subscription}:modifyOffset", SubSetOffset},
-	{"subscriptions:modifyAcl", "POST", "/projects/{project}/subscriptions/{subscription}:modifyAcl", SubModACL},
-	{"topics:list", "GET", "/projects/{project}/topics", TopicListAll},
-	{"topics:acl", "GET", "/projects/{project}/topics/{topic}:acl", TopicACL},
-	{"topics:metrics", "GET", "/projects/{project}/topics/{topic}:metrics", TopicMetrics},
-	{"topics:show", "GET", "/projects/{project}/topics/{topic}", TopicListOne},
-	{"topics:create", "PUT", "/projects/{project}/topics/{topic}", TopicCreate},
-	{"topics:delete", "DELETE", "/projects/{project}/topics/{topic}", TopicDelete},
-	{"topics:publish", "POST", "/projects/{project}/topics/{topic}:publish", TopicPublish},
-	{"topics:modifyAcl", "POST", "/projects/{project}/topics/{topic}:modifyAcl", TopicModACL},
-	{"schemas:validateMessage", "POST", "/projects/{project}/schemas/{schema}:validate", SchemaValidateMessage},
-	{"schemas:create", "POST", "/projects/{project}/schemas/{schema}", SchemaCreate},
-	{"schemas:show", "GET", "/projects/{project}/schemas/{schema}", SchemaListOne},
-	{"schemas:list", "GET", "/projects/{project}/schemas", SchemaListAll},
-	{"schemas:update", "PUT", "/projects/{project}/schemas/{schema}", SchemaUpdate},
-	{"schemas:delete", "DELETE", "/projects/{project}/schemas/{schema}", SchemaDelete},
-	{"version:list", "GET", "/version", ListVersion},
+	{"ams:metrics", "GET", "/metrics", handlers.OpMetrics},
+	{"ams:healthStatus", "GET", "/status", handlers.HealthCheck},
+	{"ams:dailyMessageAverage", "GET", "/metrics/daily-message-average", handlers.DailyMessageAverage},
+	{"users:byToken", "GET", "/users:byToken/{token}", handlers.UserListByToken},
+	{"users:byUUID", "GET", "/users:byUUID/{uuid}", handlers.UserListByUUID},
+	{"users:list", "GET", "/users", handlers.UserListAll},
+	{"users:profile", "GET", "/users/profile", handlers.UserProfile},
+	{"users:show", "GET", "/users/{user}", handlers.UserListOne},
+	{"users:refreshToken", "POST", "/users/{user}:refreshToken", handlers.RefreshToken},
+	{"users:create", "POST", "/users/{user}", handlers.UserCreate},
+	{"users:update", "PUT", "/users/{user}", handlers.UserUpdate},
+	{"users:delete", "DELETE", "/users/{user}", handlers.UserDelete},
+	{"registrations:newUser", "POST", "/registrations", handlers.RegisterUser},
+	{"registrations:acceptNewUser", "POST", "/registrations/{uuid}:accept", handlers.AcceptRegisterUser},
+	{"registrations:declineNewUser", "POST", "/registrations/{uuid}:decline", handlers.DeclineRegisterUser},
+	{"registrations:show", "GET", "/registrations/{uuid}", handlers.ListOneRegistration},
+	{"registrations:list", "GET", "/registrations", handlers.ListAllRegistrations},
+	{"projects:list", "GET", "/projects", handlers.ProjectListAll},
+	{"projects:metrics", "GET", "/projects/{project}:metrics", handlers.ProjectMetrics},
+	{"projects:addUser", "POST", "/projects/{project}/members/{user}:add", handlers.ProjectUserAdd},
+	{"projects:removeUser", "POST", "/projects/{project}/members/{user}:remove", handlers.ProjectUserRemove},
+	{"projects:showUser", "GET", "/projects/{project}/members/{user}", handlers.ProjectUserListOne},
+	{"projects:createUser", "POST", "/projects/{project}/members/{user}", handlers.ProjectUserCreate},
+	{"projects:updateUser", "PUT", "/projects/{project}/members/{user}", handlers.ProjectUserUpdate},
+	{"projects:listUsers", "GET", "/projects/{project}/members", handlers.ProjectListUsers},
+	{"projects:show", "GET", "/projects/{project}", handlers.ProjectListOne},
+	{"projects:create", "POST", "/projects/{project}", handlers.ProjectCreate},
+	{"projects:update", "PUT", "/projects/{project}", handlers.ProjectUpdate},
+	{"projects:delete", "DELETE", "/projects/{project}", handlers.ProjectDelete},
+	{"subscriptions:list", "GET", "/projects/{project}/subscriptions", handlers.SubListAll},
+	{"subscriptions:listByTopic", "GET", "/projects/{project}/topics/{topic}/subscriptions", handlers.ListSubsByTopic},
+	{"subscriptions:offsets", "GET", "/projects/{project}/subscriptions/{subscription}:offsets", handlers.SubGetOffsets},
+	{"subscriptions:timeToOffset", "GET", "/projects/{project}/subscriptions/{subscription}:timeToOffset", handlers.SubTimeToOffset},
+	{"subscriptions:acl", "GET", "/projects/{project}/subscriptions/{subscription}:acl", handlers.SubACL},
+	{"subscriptions:metrics", "GET", "/projects/{project}/subscriptions/{subscription}:metrics", handlers.SubMetrics},
+	{"subscriptions:show", "GET", "/projects/{project}/subscriptions/{subscription}", handlers.SubListOne},
+	{"subscriptions:create", "PUT", "/projects/{project}/subscriptions/{subscription}", handlers.SubCreate},
+	{"subscriptions:delete", "DELETE", "/projects/{project}/subscriptions/{subscription}", handlers.SubDelete},
+	{"subscriptions:pull", "POST", "/projects/{project}/subscriptions/{subscription}:pull", handlers.SubPull},
+	{"subscriptions:acknowledge", "POST", "/projects/{project}/subscriptions/{subscription}:acknowledge", handlers.SubAck},
+	{"subscriptions:verifyPushEndpoint", "POST", "/projects/{project}/subscriptions/{subscription}:verifyPushEndpoint", handlers.SubVerifyPushEndpoint},
+	{"subscriptions:modifyAckDeadline", "POST", "/projects/{project}/subscriptions/{subscription}:modifyAckDeadline", handlers.SubModAck},
+	{"subscriptions:modifyPushConfig", "POST", "/projects/{project}/subscriptions/{subscription}:modifyPushConfig", handlers.SubModPush},
+	{"subscriptions:modifyOffset", "POST", "/projects/{project}/subscriptions/{subscription}:modifyOffset", handlers.SubSetOffset},
+	{"subscriptions:modifyAcl", "POST", "/projects/{project}/subscriptions/{subscription}:modifyAcl", handlers.SubModACL},
+	{"topics:list", "GET", "/projects/{project}/topics", handlers.TopicListAll},
+	{"topics:acl", "GET", "/projects/{project}/topics/{topic}:acl", handlers.TopicACL},
+	{"topics:metrics", "GET", "/projects/{project}/topics/{topic}:metrics", handlers.TopicMetrics},
+	{"topics:show", "GET", "/projects/{project}/topics/{topic}", handlers.TopicListOne},
+	{"topics:create", "PUT", "/projects/{project}/topics/{topic}", handlers.TopicCreate},
+	{"topics:delete", "DELETE", "/projects/{project}/topics/{topic}", handlers.TopicDelete},
+	{"topics:publish", "POST", "/projects/{project}/topics/{topic}:publish", handlers.TopicPublish},
+	{"topics:modifyAcl", "POST", "/projects/{project}/topics/{topic}:modifyAcl", handlers.TopicModACL},
+	{"schemas:validateMessage", "POST", "/projects/{project}/schemas/{schema}:validate", handlers.SchemaValidateMessage},
+	{"schemas:create", "POST", "/projects/{project}/schemas/{schema}", handlers.SchemaCreate},
+	{"schemas:show", "GET", "/projects/{project}/schemas/{schema}", handlers.SchemaListOne},
+	{"schemas:list", "GET", "/projects/{project}/schemas", handlers.SchemaListAll},
+	{"schemas:update", "PUT", "/projects/{project}/schemas/{schema}", handlers.SchemaUpdate},
+	{"schemas:delete", "DELETE", "/projects/{project}/schemas/{schema}", handlers.SchemaDelete},
+	{"version:list", "GET", "/version", handlers.ListVersion},
 }
